refactor: #11 업적 관련 도메인 위치 변경 및 파일명 수정
