+++ RevUp-common/src/main/java/com/revup/achieve/AchieveType.java
@@ -1,9 +1,10 @@
-package com.revup.archive.entity;
+package com.revup.achieve;
 
 import lombok.RequiredArgsConstructor;
 
 @RequiredArgsConstructor
-public enum ArchiveType {
+
+public enum AchieveType {
 
     REQUESTED_REVIEW("받은 리뷰 요청 수 %d개 이상"),   //리뷰 요청
     CUMULATIVE_LIKES("누적 추천 수 %d개 이상"),   //좋아요
