+++ RevUp-domain/src/main/java/com/revup/user/service/UserReader.java
@@ -20,6 +20,7 @@ public class UserReader {
      * @param tokenInfo
      * @return
      */
+
     public User findByTokenInfo(TokenInfo tokenInfo) {
         return userAdaptor.findByTokenClaim(
                 tokenInfo.socialId(),
